requirements.txt
bson==0.5.8
certifi==2020.4.5.1
chardet==3.0.4
Django==2.1.5
django-credit-cards==0.4.1
django-crispy-forms==1.9.0
django-realtime==1.1
djongo==1.3.1
dnspython==1.16.0
docopt==0.6.2
idna==2.9
Pillow==7.1.1
pkg-resources==0.0.0
pymongo==3.10.1
python-dateutil==2.8.1
python-decouple==3.6
pytz==2019.3
requests==2.23.0
simplejson==3.17.0
six==1.14.0
sqlparse==0.2.4
urllib3==1.25.8
yarg==0.1.9
